+++ NEWS.md
@@ -1,7 +1,8 @@
 ## Changes since last release
-- sort `models` parameter in `load_forecasts()` and `load_latest_forecasts()` so that the resulting data frame is 
+- Support loading data from local zoltar module in `load_forecasts()` when `source = "local_zoltar"`. Please follow instructions in  `load_forecasts_local_zoltar()` to set up required environment for this functionality.
+- Sort `models` parameter in `load_forecasts()` and `load_latest_forecasts()` so that the resulting data frame is 
 locale-independent
-- add `hub` parameter in `get_all_models()`. It does not support loading model names for ECDC hub from remote hub repo for now. 
+- Add `hub` parameter in `get_all_models()`. It does not support loading model names for ECDC hub from remote hub repo for now. 
 
 ## covidHubUtils 0.1.6